refactor: move app handler to app.rs
